pyproject.toml
[project]
name = "grpo-math"
version = "0.1.0"
description = "Math GRPO v0 training loop (GSM8K binary reward) with DeepSpeed ZeRO-3"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "accelerate>=0.30.0",
  "datasets>=2.19.0",
  "deepspeed>=0.14.0",
  "numpy>=1.26.0",
  "pyyaml>=6.0.1",
  "tensorboard>=2.15.0",
  "torch>=2.2.0",
  "transformers>=4.41.0",
  "trl>=0.27.0",
  "tqdm>=4.66.0",
  "wandb>=0.17.0",
]

[project.optional-dependencies]
flash = [
  "flash-attn>=2.5.0; platform_system=='Linux'",
]
dev = [
  "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["grpo_math*"]
exclude = ["outputs*", "checkpoints*", "wandb*"]

